+ New feature implemented: Now it is possible to define patterns for fetching music information if there is no ID3 tag available. Currently there are the following pattern variables available:
- %ARTIST%
- %TITLE%
- %GENRE%
- %ALBUM%
- %TRACK_NO%

With the help of these variables you can define patterns like '%ARTIST%-%TITLE%' if your music file has the following structure 'Backstreet Boys - Bigger.mp3'.
There are now two different rules for detecting informations:

- FilenameDetectorRule for simple filename based detection (FilenameDetectorRule.java)
- RootSubfolderLevelDetectorRule for folder based detection (RootSubfolderLevelDetectorRule.java)

These rules can be used with the help of the DetectorRuleConfig.java class that can be defined in the config.json. The class DetectorRulesManager.java can create DetectorRule classes with the help of the DetectorRuleConfig.java.